+++ src/components/providers/ReduxProvider.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { Provider } from "react-redux";
+import { store } from "@store";
+
+interface ReduxProviderProps {
+  children: React.ReactNode;
+}
+
+export function ReduxProvider({ children }: ReduxProviderProps) {
+  return <Provider store={store}>{children}</Provider>;
+}
